+++ packages/x6-react-shape/src/extend.ts
