+++ packages/x6-react-shape/src/index.ts
@@ -2,7 +2,7 @@ export * from './shape'
 
 import { Shape } from '@antv/x6'
 import { ReactShape } from './shape'
-import './hack'
+import './extend'
 
 export function register(name: string = 'react', force: boolean = false) {
   Shape.register(name, ReactShape, force)
